+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes in 4.7
 - SQL performance: rewrite set_class_rank_mp() function in rosariosis.sql & Update.fnc.php
 - SQL remove calc_cum_gpa_mp() function & include it in t_update_mp_stats() trigger in rosariosis.sql & Update.fnc.php
 - Remove Calculate GPA program in CalcGPA.php, Menu.php, Help_en.php & rosariosis.sql
+- Do NOT remove School Period as existing Course Periods use it in Periods.php
 
 Changes in 4.6.2
 ----------------